+++ project.nix
@@ -52,12 +52,6 @@ let
     reflex-dom-pandoc =
       pkgs.haskell.lib.dontHaddock (self.callCabal2nix "reflex-dom-pandoc" sources.reflex-dom-pandoc { });
 
-    # This version is not the default in nixpkgs, yet.
-    skylighting = super.skylighting_0_10_0_2;
-    skylighting-core = super.skylighting-core_0_10_0_2;
-    # Jailbreak pandoc to work with newer skylighting
-    pandoc = doJailbreak (dontCheck super.pandoc);
-
     # Test fails on pkgsMusl
     # https://github.com/hslua/hslua/issues/67
     hslua = if disableHsLuaTests then (dontCheck super.hslua) else super.hslua;